Some corrections, but the migration of the database does not worth yet

+++ database/migrations/2024_09_19_034456_create_almacenaje_table.php
@@ -12,12 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('almacenaje', function (Blueprint $table) {
-            $table->increments("idalm");
+            $table->bigIncrements("idalm");
             $table->String("nombre");
             $table->String("descripcion");
-            $table->int("cantidad");
+            $table->integer("cantidad");
             $table->date("feCad")->nullable();
-            $table->int("disponibilidad");
+            $table->integer("disponibilidad");
             $table->String("categoria"); //Verificar
         });
     }
